requirements.txt
numpy
pandas
scikit-learn
matplotlib
jupyter
mlflow
azureml-core
azure-ai-ml
